# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StealADeal Backend API
  version: 0.1.0
  description: |
    OpenAPI contract for the current H2-backed development backend.
    This spec reflects the implemented dealer, vehicle, lead, appointment,
    and dashboard endpoints.
servers:
  - url: http://localhost:8282
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a buyer or dealer account
      operationId: registerAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive an access token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user does not have access to the requested resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    RegisterRequest:
      type: object
      required: [displayName, email, password, role]
      properties:
        displayName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          $ref: '#/components/schemas/UserRole'
        dealerId:
          type: integer
          format: int64
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, expiresAt, userId, displayName, email, role]
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        userId:
          type: integer
          format: int64
        displayName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        dealerId:
          type: integer
          format: int64
          nullable: true
    CurrentUser:
      type: object
      required: [userId, displayName, email, role]
      properties:
        userId:
          type: integer
          format: int64
        displayName:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        dealerId:
          type: integer
          format: int64
          nullable: true
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - DEALER
        - BUYER